backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.0.1
hiredis==2.2.3

# Vector DB
pgvector==0.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# OpenAI
openai==1.3.5

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-openai==0.11b0
opentelemetry-exporter-otlp-proto-http==1.21.0
prometheus-client==0.19.0

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Workers
rq==1.15.1

# Media Processing
Pillow==10.1.0

# Environment
python-dotenv==1.0.0

